Factor works if there is no triple
If there is a triple, In the recursive call of go we would
pass a proof that the orChain without the first duplication is
correct (instead of the original hypothesis). That is done by
applying dupOr to the reorder of h. But the type system is assuming
something strange about implicit arguments, so that when we use
inferType in this term it is not what we expect.